[JAX] Splitting `csrc/modules.cpp` by category (#883)
* categorized `csrc/modules.cpp`


* adapted the build tool


---------
