+++ transformer_engine/jax/csrc/extensions.h
@@ -11,6 +11,9 @@
 #include <cstddef>
 #include <cstdint>
 #include <vector>
+#include <stdexcept>
+#include <string>
+#include <iostream>
 
 #include <cuda_runtime_api.h>
 #include <pybind11/pybind11.h>
@@ -19,41 +22,54 @@
 #include <transformer_engine/fused_attn.h>
 #include <transformer_engine/transformer_engine.h>
 #include <transformer_engine/activation.h>
+#include "common/common.h"
 #include "common/util/logging.h"
+#include "utils.h"
+
+#include <cublasLt.h>
+#include <cublas_v2.h>
+#include <cuda_runtime_api.h>
+#include <cudnn.h>
 
 namespace transformer_engine {
 namespace jax {
 
 constexpr int kMaxNumDim = 8;
 
-size_t get_activation_len(NVTE_Activation_Type activation_enum);
 
+// TODO: Rename Shape to ???
 struct Shape {
     int num_dim;
     size_t dims[kMaxNumDim];
 
-    void from_vector(const std::vector<size_t> &shape) {
-        num_dim = shape.size();
-        assert(num_dim <= kMaxNumDim);
-        std::memcpy(dims, shape.data(), num_dim * sizeof(size_t));
-    }
+    void from_vector(const std::vector<size_t> &shape);
 
-    std::vector<size_t> to_vector() const {
-        assert(num_dim <= kMaxNumDim);
-        std::vector<size_t> shape(num_dim);
-        std::memcpy(shape.data(), dims, num_dim * sizeof(size_t));
-        return shape;
-    }
+    std::vector<size_t> to_vector() const;
 };
 
-enum class NVTE_Activation_Enum {
-  GELU,
-  GEGLU,
-  SILU,
-  SWIGLU,
-};
+// Phuong: These 3 functions need to stay in the header file for compilation purpose
+// 1.
+inline bool use_fp8(DType type) {
+  return type == DType::kFloat8E4M3 || type == DType::kFloat8E5M2;
+}
+// 2.
+template <typename T>
+pybind11::bytes PackOpaque(const T &descriptor) {
+    auto str = std::string(reinterpret_cast<const char *>(&descriptor), sizeof(T));
+    return pybind11::bytes(str);
+}
+// 3.
+template <typename T>
+const T *UnpackOpaque(const char *opaque, size_t opaque_len) {
+    if (opaque_len != sizeof(T)) {
+        throw std::runtime_error("Invalid opaque object size");
+    }
+    return reinterpret_cast<const T *>(opaque);
+}
 
-size_t get_activation_len(NVTE_Activation_Enum act_enum);
+std::vector<size_t> MakeShapeVector(NVTEShape shape);
+
+// Packing
 
 struct CustomCallCommonDescriptor {
     Shape shape;
@@ -144,16 +160,21 @@ pybind11::bytes PackCustomCallFusedAttnDescriptor(
     NVTE_Mask_Type mask_type, NVTE_QKV_Layout qkv_layout, DType dtype, DType wkspace_dtype,
     bool is_training);
 
-NVTE_Fused_Attn_Backend GetFusedAttnBackend(DType q_dtype, DType kv_dtype,
-                                            NVTE_QKV_Layout qkv_layout, NVTE_Bias_Type bias_type,
-                                            NVTE_Mask_Type mask_type, float dropout_probability,
-                                            size_t q_num_heads, size_t kv_num_heads,
-                                            size_t q_max_seqlen, size_t kv_max_seqlen,
-                                            size_t head_dim);
+// Transpose
 
 void Transpose(cudaStream_t stream, void **buffers, const char *opaque, size_t opaque_len);
 
 void CastTranspose(cudaStream_t stream, void **buffers, const char *opaque, size_t opaque_len);
+
+pybind11::tuple GetDBiasCastTransposeWorkspaceSizes(size_t batch_size, size_t hidden_size,
+                                                         DType in_dtype, DType out_dtype);
+
+void DBiasCastTranspose(cudaStream_t stream, void **buffers, const char *opaque,
+                             size_t opaque_len);
+
+// Activation
+
+size_t get_activation_len(NVTE_Activation_Type activation_enum);
 
 void ActLu(cudaStream_t stream, void **buffers, const char *opaque, size_t opaque_len);
 
@@ -167,14 +188,10 @@ pybind11::tuple GetDActDBiasCastTransposeWorkspaceSizes(size_t batch_size, size_
 void DActLuDBiasCastTranspose(cudaStream_t stream, void **buffers, const char *opaque,
                              size_t opaque_len);
 
-pybind11::tuple GetDBiasCastTransposeWorkspaceSizes(size_t batch_size, size_t hidden_size,
-                                                         DType in_dtype, DType out_dtype);
-
-void DBiasCastTranspose(cudaStream_t stream, void **buffers, const char *opaque,
-                             size_t opaque_len);
-
 void DGatedActLuCastTranspose(cudaStream_t stream, void **buffers, const char *opaque,
                              size_t opaque_len);
+
+// Normalization
 
 pybind11::tuple GetLayerNormForwardWorkspaceSizes(size_t batch_size, size_t hidden_size,
                                                   DType in_dtype, DType w_dtype, DType out_dtype,
@@ -199,9 +216,13 @@ void RMSNormForwardFP8(cudaStream_t stream, void **buffers, const char *opaque, 
 
 void RMSNormBackward(cudaStream_t stream, void **buffers, const char *opaque, size_t opaque_len);
 
+// Quantization
+
 void Quantize(cudaStream_t stream, void **buffers, const char *opaque, size_t opaque_len);
 
 void Dequantize(cudaStream_t stream, void **buffers, const char *opaque, size_t opaque_len);
+
+// Softmax
 
 void ScaledSoftmaxForward(cudaStream_t stream, void **buffers, const char *opaque,
                           std::size_t opaque_len);
@@ -220,6 +241,15 @@ void ScaledUpperTriangMaskedSoftmaxForward(cudaStream_t stream, void **buffers, 
 
 void ScaledUpperTriangMaskedSoftmaxBackward(cudaStream_t stream, void **buffers, const char *opaque,
                                             std::size_t opaque_len);
+
+// Attention
+
+NVTE_Fused_Attn_Backend GetFusedAttnBackend(DType q_dtype, DType kv_dtype,
+                                            NVTE_QKV_Layout qkv_layout, NVTE_Bias_Type bias_type,
+                                            NVTE_Mask_Type mask_type, float dropout_probability,
+                                            size_t q_num_heads, size_t kv_num_heads,
+                                            size_t q_max_seqlen, size_t kv_max_seqlen,
+                                            size_t head_dim);
 
 pybind11::tuple GetFusedAttnForwardWorkspaceSizes(
     size_t input_batch, size_t bias_batch, size_t q_max_seqlen, size_t kv_max_seqlen,
